+++ modules/AWSPowerShell/Cmdlets/IoTSiteWise/AWS.Tools.IoTSiteWise.Completers.psm1
@@ -118,6 +118,20 @@ $IOTSW_Completers = {
             break
         }
 
+        # Amazon.IoTSiteWise.ListAssetModelPropertiesFilter
+        "Get-IOTSWAssetModelPropertyList/Filter"
+        {
+            $v = "ALL","BASE"
+            break
+        }
+
+        # Amazon.IoTSiteWise.ListAssetPropertiesFilter
+        "Get-IOTSWAssetPropertyList/Filter"
+        {
+            $v = "ALL","BASE"
+            break
+        }
+
         # Amazon.IoTSiteWise.ListAssetsFilter
         "Get-IOTSWAssetList/Filter"
         {
@@ -220,7 +234,7 @@ $IOTSW_map = @{
     "AccessPolicyPermission"=@("New-IOTSWAccessPolicy","Update-IOTSWAccessPolicy")
     "DisassociatedDataStorage"=@("Write-IOTSWStorageConfiguration")
     "EncryptionType"=@("Write-IOTSWDefaultEncryptionConfiguration")
-    "Filter"=@("Get-IOTSWAssetList","Get-IOTSWBulkImportJobList")
+    "Filter"=@("Get-IOTSWAssetList","Get-IOTSWAssetModelPropertyList","Get-IOTSWAssetPropertyList","Get-IOTSWBulkImportJobList")
     "IdentityType"=@("Get-IOTSWAccessPolicyList")
     "LoggingOptions_Level"=@("Write-IOTSWLoggingOption")
     "PortalAuthMode"=@("New-IOTSWPortal")
@@ -331,7 +345,9 @@ $IOTSW_SelectMap = @{
                "Get-IOTSWAssetPropertyValueHistory",
                "Get-IOTSWInterpolatedAssetPropertyValue",
                "Get-IOTSWAccessPolicyList",
+               "Get-IOTSWAssetModelPropertyList",
                "Get-IOTSWAssetModelList",
+               "Get-IOTSWAssetPropertyList",
                "Get-IOTSWAssetRelationshipList",
                "Get-IOTSWAssetList",
                "Get-IOTSWAssociatedAssetList",