+++ modules/AWSPowerShell/Cmdlets/SSMIncidents/Basic/New-SSMIReplicationSet-Cmdlet.cs
@@ -59,10 +59,21 @@ namespace Amazon.PowerShell.Cmdlets.SSMI
         public System.Collections.Hashtable ReplicationSetRegions { get; set; }
         #endregion
         
+        #region Parameter Tag
+        /// <summary>
+        /// <para>
+        /// <para>A list of tags to add to the replication set.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [Alias("Tags")]
+        public System.Collections.Hashtable Tag { get; set; }
+        #endregion
+        
         #region Parameter ClientToken
         /// <summary>
         /// <para>
-        /// <para>A token ensuring that the operation is called only once with the specified details.</para>
+        /// <para>A token that ensures that the operation is called only once with the specified details.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
@@ -125,6 +136,14 @@ namespace Amazon.PowerShell.Cmdlets.SSMI
                 WriteWarning("You are passing $null as a value for parameter ReplicationSetRegions which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
             }
             #endif
+            if (this.Tag != null)
+            {
+                context.Tag = new Dictionary<System.String, System.String>(StringComparer.Ordinal);
+                foreach (var hashKey in this.Tag.Keys)
+                {
+                    context.Tag.Add((String)hashKey, (String)(this.Tag[hashKey]));
+                }
+            }
             
             // allow further manipulation of loaded context prior to processing
             PostExecutionContextLoad(context);
@@ -148,6 +167,10 @@ namespace Amazon.PowerShell.Cmdlets.SSMI
             if (cmdletContext.ReplicationSetRegions != null)
             {
                 request.Regions = cmdletContext.ReplicationSetRegions;
+            }
+            if (cmdletContext.Tag != null)
+            {
+                request.Tags = cmdletContext.Tag;
             }
             
             CmdletOutput output;
@@ -212,6 +235,7 @@ namespace Amazon.PowerShell.Cmdlets.SSMI
         {
             public System.String ClientToken { get; set; }
             public Dictionary<System.String, Amazon.SSMIncidents.Model.RegionMapInputValue> ReplicationSetRegions { get; set; }
+            public Dictionary<System.String, System.String> Tag { get; set; }
             public System.Func<Amazon.SSMIncidents.Model.CreateReplicationSetResponse, NewSSMIReplicationSetCmdlet, object> Select { get; set; } =
                 (response, cmdlet) => response.Arn;
         }